requirements.txt
pandas
bcrypt
streamlit
plotly
google-generativeai
pillow